Create basic skeleton
This contains README, LICENSE, embryonic manual and a
preliminary version of the "binary file" that serves as the
main program. It can only parse program arguments and show
help and version but nothing more. The documentation files
deal with generic topics like the coding standards and the
story behind this project as there is yet no actual
functionality to be documented.

Additionally, as the manual says, any new code in the project
must be exercised by at least one test case. So this also
adds a test suite with two test cases exercising the two
simple options already implemented in the main program.