Add installation and resource checks
Add the main LUA script which for now just checks that the
proper version of LUA is used to run the program. Also add
into the main executable code to locate the Lua resources and
sections of the manual and README.rst file about how to
install the program.

For simplicity, the new code in the main executable simply
runs the Lua portion of the program once found, taking care
to stay in whatever directory the executable was invoked.
The main executable still has some more tasks to do though,
which will be implemented in later commits.